refactor(core): fold document renderer into converter
