+++ packages/core/src/converter.ts
@@ -1,10 +1,36 @@
+import { readFileSync, statSync } from "node:fs";
+
+import { createComputedView, readPptx } from "@pptx-glimpse/document";
 import type { FontMapping } from "@pptx-glimpse/renderer";
 import type { LogLevel } from "@pptx-glimpse/renderer";
+import {
+  buildFontFaceStyle,
+  collectFontFilePaths,
+  createFontMapping,
+  createOpentypeSetupFromSystem,
+  DEFAULT_OUTPUT_WIDTH,
+  flushWarnings,
+  FontUsageCollector,
+  initWarningLogger,
+  renderSlideToSvg,
+  resetFontMapping,
+  resetFontUsageCollector,
+  resetScriptFonts,
+  resetTextMeasurer,
+  resetTextPathFontResolver,
+  setFontMapping,
+  setFontUsageCollector,
+  setScriptFonts,
+  setTextMeasurer,
+  setTextPathFontResolver,
+  svgToPng,
+  warn,
+} from "@pptx-glimpse/renderer";
 
 import {
-  convertPptxToPngViaDocumentPath,
-  convertPptxToSvgViaDocumentPath,
-} from "./document-path-renderer.js";
+  adaptComputedViewToRendererModel,
+  type RendererAdapterDiagnostic,
+} from "./pptx-computed-view-renderer-adapter.js";
 
 export interface ConvertOptions {
   /** 変換対象のスライド番号 (1始まり)。未指定で全スライド */
@@ -44,11 +70,23 @@ export interface SlideImage {
   height: number;
 }
 
+type ConversionDiagnostic = RendererAdapterDiagnostic;
+
+interface SvgConversionResult {
+  readonly slides: readonly SlideSvg[];
+  readonly diagnostics: readonly ConversionDiagnostic[];
+}
+
+interface PngConversionResult {
+  readonly slides: readonly SlideImage[];
+  readonly diagnostics: readonly ConversionDiagnostic[];
+}
+
 export async function convertPptxToSvg(
   input: Buffer | Uint8Array,
   options?: ConvertOptions,
 ): Promise<SlideSvg[]> {
-  const result = await convertPptxToSvgViaDocumentPath(input, options);
+  const result = await convertPptxToSvgInternal(input, options);
   return [...result.slides];
 }
 
@@ -56,6 +94,159 @@ export async function convertPptxToPng(
   input: Buffer | Uint8Array,
   options?: ConvertOptions,
 ): Promise<SlideImage[]> {
-  const result = await convertPptxToPngViaDocumentPath(input, options);
+  const result = await convertPptxToPngInternal(input, options);
   return [...result.slides];
+}
+
+async function convertPptxToSvgInternal(
+  input: Buffer | Uint8Array,
+  options?: ConvertOptions,
+): Promise<SvgConversionResult> {
+  const textOutput = options?.textOutput ?? "path";
+  const setup = await createOpentypeSetupFromSystem(
+    options?.fontDirs,
+    options?.fontMapping,
+    options?.skipSystemFonts,
+  );
+  if (setup) {
+    setTextMeasurer(setup.measurer);
+    if (textOutput !== "text") {
+      setTextPathFontResolver(setup.fontResolver);
+    }
+  }
+
+  const fontUsageCollector = textOutput === "text" ? new FontUsageCollector() : null;
+  if (fontUsageCollector) {
+    setFontUsageCollector(fontUsageCollector);
+  }
+  setFontMapping(createFontMapping(options?.fontMapping));
+
+  try {
+    initWarningLogger(options?.logLevel ?? "off");
+
+    const source = readPptx(input);
+    const scriptFontScheme = findScriptFontScheme(source);
+    setScriptFonts(
+      scriptFontScheme?.majorJapanese ?? null,
+      scriptFontScheme?.minorJapanese ?? null,
+    );
+    if (source.presentation.slidePartPaths.length === 0) {
+      warn("presentation.noSlides", "No slides found in the PPTX file");
+    }
+
+    const computed = createComputedView(source, { slides: options?.slides });
+    const adapted = adaptComputedViewToRendererModel(computed);
+    const diagnostics = adapted.diagnostics;
+    const slideSize = adapted.slideSize;
+    if (slideSize === undefined && adapted.slides.length > 0) {
+      throw new Error("Converter requires a computed slide size");
+    }
+
+    const slides: SlideSvg[] = [];
+    for (const slide of adapted.slides) {
+      if (slideSize === undefined) continue;
+      fontUsageCollector?.reset();
+      let svg = renderSlideToSvg(slide, slideSize);
+      if (fontUsageCollector && setup) {
+        const style = await buildFontFaceStyle(fontUsageCollector.getUsages(), setup.fontResolver);
+        if (style) {
+          svg = injectIntoSvgDefs(svg, style);
+        }
+      }
+      slides.push({ slideNumber: slide.slideNumber, svg });
+    }
+
+    flushWarnings();
+    return { slides, diagnostics };
+  } finally {
+    resetTextMeasurer();
+    resetTextPathFontResolver();
+    resetFontUsageCollector();
+    resetFontMapping();
+    resetScriptFonts();
+  }
+}
+
+function findScriptFontScheme(source: ReturnType<typeof readPptx>) {
+  const firstThemePartPath = source.slideMasters.find(
+    (master) => master.themePartPath !== undefined,
+  )?.themePartPath;
+  return (
+    source.themes.find((theme) => theme.partPath === firstThemePartPath)?.fontScheme ??
+    source.themes[0]?.fontScheme
+  );
+}
+
+async function convertPptxToPngInternal(
+  input: Buffer | Uint8Array,
+  options?: ConvertOptions,
+): Promise<PngConversionResult> {
+  const svgResult = await convertPptxToSvgInternal(input, {
+    ...options,
+    textOutput: "path",
+  });
+  const width = options?.width ?? DEFAULT_OUTPUT_WIDTH;
+  const height = options?.height;
+  const fontBuffers = loadFontBuffers(options?.fontDirs, options?.skipSystemFonts);
+
+  const slides: SlideImage[] = [];
+  for (const { slideNumber, svg } of svgResult.slides) {
+    const pngResult = await svgToPng(svg, { width, height, fontBuffers });
+    slides.push({
+      slideNumber,
+      png: pngResult.png,
+      width: pngResult.width,
+      height: pngResult.height,
+    });
+  }
+
+  return { slides, diagnostics: svgResult.diagnostics };
+}
+
+function injectIntoSvgDefs(svg: string, content: string): string {
+  const openTagEnd = svg.indexOf(">");
+  if (openTagEnd === -1) return svg;
+  return `${svg.slice(0, openTagEnd + 1)}<defs>${content}</defs>${svg.slice(openTagEnd + 1)}`;
+}
+
+let cachedFontBuffers: Uint8Array[] | null = null;
+let cachedFontBuffersKey: string | null = null;
+
+const MAX_TOTAL_FONT_BUFFER_BYTES = 100 * 1024 * 1024;
+
+function loadFontBuffers(fontDirs?: string[], skipSystemFonts?: boolean): Uint8Array[] {
+  const key = `${(fontDirs ?? []).join("\0")}\n${skipSystemFonts ?? false}`;
+  if (cachedFontBuffers !== null && cachedFontBuffersKey === key) {
+    return cachedFontBuffers;
+  }
+
+  const fontPaths = collectFontFilePaths(fontDirs, skipSystemFonts).filter((path) => {
+    const lower = path.toLowerCase();
+    return lower.endsWith(".ttf") || lower.endsWith(".otf");
+  });
+  const readableFontPaths: { path: string; size: number }[] = [];
+  for (const path of fontPaths) {
+    try {
+      readableFontPaths.push({ path, size: statSync(path).size });
+    } catch {
+      // Ignore unreadable font files.
+    }
+  }
+  readableFontPaths.sort((a, b) => a.size - b.size || a.path.localeCompare(b.path));
+
+  const buffers: Uint8Array[] = [];
+  let totalSize = 0;
+  for (const { path, size } of readableFontPaths) {
+    if (totalSize + size > MAX_TOTAL_FONT_BUFFER_BYTES) break;
+    try {
+      buffers.push(new Uint8Array(readFileSync(path)));
+      totalSize += size;
+    } catch {
+      // Ignore fonts that disappear between stat and read.
+    }
+  }
+
+  cachedFontBuffers = buffers;
+  cachedFontBuffersKey = key;
+  return buffers;
 }
